requirements.txt
CacheControl
Jinja2
PyYAML
accelerate
charset-normalizer
clikit
cycler
diffusers
einops
importlib_resources
ipywidgets
lockfile
matplotlib
numpy
opencv-python
packaging
pandas
pillow
safetensors
scikit-learn
scipy
torch
torchmetrics
torchvision
tqdm
transformers
triton
urllib3
xformer
yarl
